// src/app/models/studentaddress.ts
export class StudentAddress {
  studentAddressId: number;
  studentId: number;
  address1: string;
  address2: string;
  city: string;
  subCountry: string;
  stateCode: string;
  provinceCode: string;
  countryCode: string;
  zip: string;
  primaryContactPhoneNumber: string;
  secondaryContactPhoneNumber: string;
  primaryContactEmailAddress: string;
  secondaryContactEmailAddress: string;
  addressTypeCode: string;
  createTime: string;
}