// src/app/models/attendancesummary.ts
export class AttendanceSummary  {
  numberOfPresentDays: number;
  numberOfExcusedDays: number;
  numberOfUnexcusedDays: number;
  numberOfSuspendedDays: number;
  numberOfTardyDays: number;
  numberOfTruancyDays: number;
  presentDays: string[];
  excusedDays: string[];
  unexcusedDays: string[];
  suspendedDays: string[];
  tardyDays: string[];
  truancyDays: string[];
}